+++ CHANGES.md
@@ -4,7 +4,8 @@
 
 * Update sentinel2-cogs image path (https://github.com/cogeotiff/rio-tiler-pds/pull/22).
 * Remove ContextManager requirement in base class and update for rio-tiler 2.0b13 (https://github.com/cogeotiff/rio-tiler/pull/265).
-* Add MODIS (PDS and Astraea) dataset ()
+* Add MODIS (PDS and Astraea) dataset (https://github.com/cogeotiff/rio-tiler-pds/issues/18)
+* move reader base classes to rio-tiler (https://github.com/cogeotiff/rio-tiler-pds/issues/24)
 
 ## 0.2.1 (2020-09-25)